deltahacks/requirements.txt

bcrypt==3.1.7
firebase-admin==2.17.0
Flask==1.0.2
gcloud==0.18.3
google-api-core==1.12.0
google-api-python-client==1.7.8
oauth2client==4.1.3
oauthlib==3.0.1
Pillow==7.0.0
pytesseract==0.3.2
